\name{Q2}
\alias{Q2}
\title{
Quantile2
}
\description{
Quantile 2
}
\usage{
Q2(par, tau, j, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{par}{Matrix of parameters (nmeric 4xK - matrix); may also be delivered separately.}
  \item{tau}{A value}
   \item{j}{Possible values (numeric vector of length J > 0)}
   \item{n}{Frequencies (integer vector of length J; non-negative entries)}
}

\author{
Lisa Allmesberger
Fabian Bergs
Stefan Immler
Michael Kässmann
}
