\name{Q3}
\alias{Q3}
\title{
Quantile3
}
\description{
Quantile 3
}
\usage{
Q3(s, a, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s}{a}
  \item{a}{a}
  \item{b}{a}
}

\author{
Lisa Allmesberger
Fabian Bergs
Stefan Immler
Michael Kässmann
}
